+++ app/src/main/java/hexlet/code/controller/UrlController.java
@@ -85,7 +85,7 @@ public class UrlController {
     }
 
     public static void show(Context ctx) throws SQLException {
-        var id = ctx.pathParamAsClass("id", int.class).get();
+        var id = ctx.pathParamAsClass("id", Long.class).get();
         var url = UrlRepository.findById(id)
                 .orElseThrow(() -> new NotFoundResponse("Url not found"));
         List<UrlCheck> urlChecks = new ArrayList<>();
