+++ app/src/main/java/hexlet/code/model/UrlCheck.java
@@ -8,15 +8,15 @@ import java.sql.Timestamp;
 @Getter
 @Setter
 public final class UrlCheck {
-    private int id;
-    private int urlId;
+    private Long id;
+    private Long urlId;
     private int statusCode;
     private String h1;
     private String title;
     private String description;
     private Timestamp createdAt;
 
-    public UrlCheck(int urlId, int statusCode, String h1, String title, String description, Timestamp createdAt) {
+    public UrlCheck(Long urlId, int statusCode, String h1, String title, String description, Timestamp createdAt) {
         this.urlId = urlId;
         this.statusCode = statusCode;
         this.h1 = h1;
