+++ app/src/main/java/hexlet/code/util/NamedRoutes.java
@@ -8,7 +8,7 @@ public class NamedRoutes {
     public static String urlsPath() {
         return "/urls";
     }
-    public static String urlPath(int id) {
+    public static String urlPath(Long id) {
         return urlPath(String.valueOf(id));
     }
 
@@ -16,7 +16,7 @@ public class NamedRoutes {
         return "/urls/" + id;
     }
 
-    public static String urlCheckPath(int id) {
+    public static String urlCheckPath(Long id) {
         return urlCheckPath(String.valueOf(id));
     }
